Edited some files before deploying

+++ src/components/Nweet.js
@@ -11,7 +11,7 @@ const Nweet = ({nweetObj, isOwner}) => {
         const ok = window.confirm("Are you sure you want to delete this nweet?");
         if(ok) {
             await dbService.doc(`nweets/${nweetObj.id}`).delete();
-            if(nweetObj.attachmentUrl !== null) { 
+            if(nweetObj.attachmentUrl !== null && nweetObj.attachmentUrl !== '') { 
                 await storageService.refFromURL(nweetObj.attachmentUrl).delete();
             }
         }
